# Make finding corner neighbor symmetrical in ChessBoardDetector::findQuadNeighbors