--- a/modules/calib3d/src/calibinit.cpp
+++ b/modules/calib3d/src/calibinit.cpp
@@ -227,7 +227,7 @@ public:
         std::vector<int> neighbors_indices;
         std::vector<float> neighbors_dists;
         std::vector<Point2f> all_quads_pts;
-        Ptr<flann::GenericIndex<flann::L2_Simple<float>>> p_all_quads_pts_index;
+        flann::GenericIndex<flann::L2_Simple<float>> all_quads_pts_index;
 
         NeighborsFinder(ChessBoardDetector& detector);
 
@@ -491,20 +491,27 @@ static void icvBinarizationHistogramBased(Mat & img)
     }
 }
 
-ChessBoardDetector::NeighborsFinder::NeighborsFinder(ChessBoardDetector& _detector): detector(_detector) {
-    const int all_corners_count = detector.all_quads_count * 4;
-    neighbors_indices.resize(all_corners_count);
-    neighbors_dists.resize(all_corners_count);
-    all_quads_pts.reserve(all_corners_count);
-    for (int idx = 0; idx < detector.all_quads_count; idx++)
+static std::vector<Point2f> getCornersFromQuads(ChessBoardQuad* p_all_quads, const int all_quads_count)
+{
+    std::vector<Point2f> all_quads_pts;
+    all_quads_pts.reserve(all_quads_count * 4);
+    for (int idx = 0; idx < all_quads_count; idx++)
     {
-        const ChessBoardQuad& cur_quad = (const ChessBoardQuad&)detector.all_quads[idx];
+        const ChessBoardQuad& cur_quad = (const ChessBoardQuad&)p_all_quads[idx];
         for (int i = 0; i < 4; i++)
             all_quads_pts.push_back(cur_quad.corners[i]->pt);
     }
-    const cvflann::KDTreeSingleIndexParams index_params;
-    p_all_quads_pts_index = makePtr<flann::GenericIndex<flann::L2_Simple<float>>>
-                            (Mat(all_quads_pts).reshape(1, all_corners_count), index_params);
+    return all_quads_pts;
+}
+
+ChessBoardDetector::NeighborsFinder::NeighborsFinder(ChessBoardDetector& _detector) :
+    detector(_detector),
+    all_quads_pts(getCornersFromQuads(detector.all_quads.data(), detector.all_quads_count)),
+    all_quads_pts_index(Mat(all_quads_pts).reshape(1, detector.all_quads_count * 4), cvflann::KDTreeSingleIndexParams())
+{
+    const int all_corners_count = detector.all_quads_count * 4;
+    neighbors_indices.resize(all_corners_count);
+    neighbors_dists.resize(all_corners_count);
 }
 
 bool ChessBoardDetector::NeighborsFinder::findCornerNeighbor(
@@ -517,14 +524,13 @@ bool ChessBoardDetector::NeighborsFinder::findCornerNeighbor(
     cv::Point2f& closest_corner_pt)
 {
     ChessBoardQuad* p_all_quads = detector.all_quads.data();
-    flann::GenericIndex<flann::L2_Simple<float>>& all_quads_pts_index = *p_all_quads_pts_index;
 
     const ChessBoardQuad& cur_quad = (const ChessBoardQuad&)p_all_quads[idx];
     int closest_neighbor_idx = -1;
     ChessBoardQuad *closest_quad = 0;
 
     // find the closest corner in all other quadrangles
-    const std::vector<float> query = Mat(pt);
+    const std::vector<float> query = { pt.x, pt.y };
     const cvflann::SearchParams search_params(-1);
     const int neighbors_count = all_quads_pts_index.radiusSearch(query, neighbors_indices, neighbors_dists, radius, search_params);