@AleksandrPanov 2024-06-03
@MaximSmolskiy, it looks like there are too many parameters in `findCornerNeighbor()`:
```
static bool findCornerNeighbor(
    ChessBoardQuad* all_quads,
    const std::vector<Point2f>& all_quads_pts,
    flann::GenericIndex<flann::L2_Simple<float>>& all_quads_pts_index,
    const int idx,
    const cv::Point2f& pt,
    float& min_dist,
    const float radius,
    std::vector<int>& neighbors_indices,
    std::vector<float>& neighbors_dists,
    int& closest_quad_idx,
    int& closest_corner_idx,
    cv::Point2f& closest_corner_pt)
```
I added the `NeighborsFinder` struct and simplified the API with it:
```
bool ChessBoardDetector::NeighborsFinder::findCornerNeighbor(
    const int idx,
    const cv::Point2f& pt,
    float& min_dist,
    const float radius,
    int& closest_quad_idx,
    int& closest_corner_idx,
    cv::Point2f& closest_corner_pt)
```




Do you have any comments or suggestions?

Otherwise, I like it.