@MaximSmolskiy 2024-06-04
@AleksandrPanov Thanks for refactoring! I don't have any suggestions